test(licenses): reorganize integration test files
