+++ server/tests/integration/licenses/catalog-update/license-billing-mode-validation.test.ts
@@ -1,6 +1,4 @@
 import { test } from "bun:test";
-import type { AttachParamsV1Input } from "@autumn/shared";
-import { TestFeature } from "@tests/setup/v2Features.js";
 import { expectAutumnError } from "@tests/utils/expectUtils/expectErrUtils.js";
 import { items } from "@tests/utils/fixtures/items.js";
 import { products } from "@tests/utils/fixtures/products.js";
@@ -8,22 +6,21 @@ import { initScenario, s } from "@tests/utils/testInitUtils/initScenario.js";
 import chalk from "chalk";
 
 test.concurrent(
-	`${chalk.yellowBright("licenses overflow blocked: attach add_licenses with prepaid_only false rejects")}`,
+	`${chalk.yellowBright("licenses overflow blocked: link with prepaid_only false rejects")}`,
 	async () => {
 		const parent = products.base({
-			id: "ovf-block-attach-parent",
+			id: "ovf-block-parent",
 			items: [items.dashboard()],
 		});
 		const license = products.base({
-			id: "ovf-block-attach-seat",
+			id: "ovf-block-seat",
 			items: [items.monthlyMessages({ includedUsage: 25 })],
 		});
 
-		const { customerId, autumnV2_2 } = await initScenario({
-			customerId: "license-overflow-attach-blocked",
+		const { autumnV2_2 } = await initScenario({
+			customerId: "license-overflow-blocked",
 			setup: [
 				s.customer({ testClock: false }),
-				s.entities({ count: 1, featureId: TestFeature.Users }),
 				s.products({ list: [parent, license] }),
 			],
 			actions: [],
@@ -32,18 +29,11 @@ test.concurrent(
 		await expectAutumnError({
 			errMessage: "not yet available",
 			func: () =>
-				autumnV2_2.billing.attach<AttachParamsV1Input>({
-					customer_id: customerId,
+				autumnV2_2.post("/plans.update", {
 					plan_id: parent.id,
-					customize: {
-						add_licenses: [
-							{
-								license_plan_id: license.id,
-								included: 1,
-								prepaid_only: false,
-							},
-						],
-					},
+					licenses: [
+						{ license_plan_id: license.id, included: 1, prepaid_only: false },
+					],
 				}),
 		});
 	},
