+++ server/tests/integration/licenses/catalog-update/license-sandbox-copy.test.ts
@@ -5,11 +5,11 @@ import {
 	ErrCode,
 	FeatureUsageType,
 	mapToProductV2,
-	ResetInterval,
 	type Organization,
 	type OrgConfig,
 	organizations,
 	RecaseError,
+	ResetInterval,
 } from "@autumn/shared";
 import { products } from "@tests/utils/fixtures/products.js";
 import defaultCtx from "@tests/utils/testInitUtils/createTestContext.js";
@@ -24,11 +24,12 @@ import {
 	constructCreditSystem,
 	constructMeteredFeature,
 } from "@/internal/features/utils/constructFeatureUtils.js";
+import { loadApiPlanLicenses } from "@/internal/licenses/actions/links/loadApiPlanLicenses.js";
+import { syncPlanLicenses } from "@/internal/licenses/actions/links/syncPlanLicenses.js";
+import { planLicenseRepo } from "@/internal/licenses/repos/planLicenseRepo.js";
 import { deletePlatformSubOrg } from "@/internal/orgs/deleteOrg/deletePlatformSubOrg.js";
 import { OrgService } from "@/internal/orgs/OrgService.js";
 import { createProduct } from "@/internal/product/actions/createProduct.js";
-import { linkLicense } from "@/internal/licenses/actions/links/linkLicense.js";
-import { listLicenseLinks } from "@/internal/licenses/actions/links/listLicenseLinks.js";
 import { ProductService } from "@/internal/products/ProductService.js";
 import { copySandboxForOrg } from "@/internal/sandboxes/copySandbox.js";
 import { generatePublishableKey } from "@/utils/encryptUtils.js";
@@ -181,24 +182,32 @@ const seedSourceSandbox = async (sourceOrg: Organization) => {
 			],
 		}) as unknown as CreateProductV2Params,
 	});
-	await linkLicense({
+	const parentProduct = await ProductService.getFull({
+		db,
+		orgId: sourceOrg.id,
+		env: AppEnv.Sandbox,
+		idOrInternalId: PRODUCT_ID,
+	});
+	await syncPlanLicenses({
 		ctx: { ...seedCtx, features: allFeatures },
-		params: {
-			parent_plan_id: PRODUCT_ID,
-			license_plan_id: LICENSE_ID,
-			included: 3,
-			prepaid_only: true,
-			metadata: { source: "sandbox-copy" },
-			customize: {
-				items: [
-					{
-						feature_id: MSG_FEATURE,
-						included: 80,
-						reset: { interval: ResetInterval.Month },
-					},
-				],
+		parentProduct,
+		licenses: [
+			{
+				license_plan_id: LICENSE_ID,
+				included: 3,
+				prepaid_only: true,
+				metadata: { source: "sandbox-copy" },
+				customize: {
+					items: [
+						{
+							feature_id: MSG_FEATURE,
+							included: 80,
+							reset: { interval: ResetInterval.Month },
+						},
+					],
+				},
 			},
-		},
+		],
 	});
 };
 
@@ -283,26 +292,31 @@ describe("sandboxes.copy: copy plans + features between two sandbox sub-orgs", (
 			.filter((id): id is string => Boolean(id));
 		expect(itemFeatureIds.sort()).toEqual([DASH_FEATURE, MSG_FEATURE].sort());
 
-		const links = await listLicenseLinks({
+		const linksByParent = await loadApiPlanLicenses({
 			ctx: {
 				...baseCtx,
 				org: target,
 				env: AppEnv.Sandbox,
 				features: targetFeatures,
 			},
-			parentPlanId: PRODUCT_ID,
+			internalProductIds: [copied!.internal_id],
 		});
+		const links = linksByParent.get(copied!.internal_id) ?? [];
 		expect(links).toHaveLength(1);
 		expect(links[0]).toMatchObject({
-			parent_plan_id: PRODUCT_ID,
 			license_plan_id: LICENSE_ID,
 			included: 3,
-			metadata: { source: "sandbox-copy" },
 		});
 		expect(links[0].customize?.add_items?.[0]).toMatchObject({
 			feature_id: MSG_FEATURE,
 			included: 80,
 		});
+		const [storedLink] =
+			await planLicenseRepo.listCatalogByParentInternalProductIds({
+				db,
+				parentInternalProductIds: [copied!.internal_id],
+			});
+		expect(storedLink.metadata).toEqual({ source: "sandbox-copy" });
 	}, 180_000);
 
 	test("rejects a non-owned source with a 404 (ownership masked)", async () => {
